spider/pyproject.toml
[tool.poetry]
name = "covidwire-spider"
version = "0.1.0"
description = ""
authors = ["Kevin Zeidler <kzeidler@gmail.com>"]
packages = [
    {"include" = "spider"}
]

[tool.poetry.dependencies]
python = "^3.7"
date-guesser = "^2.1.4"
simpletransformers = "^0.34.4"
httpx = "*"
bs4 = "^0.0.1"
ftfy = "^5.7"
dateutils = "^0.6.8"
gemeinsprache = "^0.1.0"
psycopg2 = "^2.8.5"
trio = "^0.16.0"
torch = "^1.5.0"
dataset = "^1.3.1"
termcolor = "^1.1.0"
pytz = "^2020.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
